Tests are now called from within measured as functions rather than executed as binaries. To maintain the ability to test them standalone each test is compiled against a small wrapper, providing a main function and argument parsing to make everything the way the test functions expect.
Destinations are passed in to the test functions as a list of addrinfo, regardless of the source being measured or the standalone test. 

There is no longer a need for tests to provide set up functions that do pre-test setup, this should now be done within the test (using functions from libmeasured that are yet to be written).