Report test results to the most appropriate place.
Tests should now include a save callback and a print callback function and the appropriate one is called depending on whether the test is being run from measured or standalone. A few dummy functions are provided in testmain.c to replicate behaviour that is not available when running standalone (e.g. looking up ampnames).