Create a new connection to the broker for each test.
Each connection can only be operated on by a single thread/process at a time, and the current approach to creating individual channels allowed multiple processes to be using that one connection. This caused data to be lost. Ideally there would only be a single connection and a channel per process, but it is awkward to create the channel in the main process and give it to the child.