Remove unnecessary nametable structure instead use a list of addrinfo structs.
To maintain all the previous behavior the ai_canonname field is being overwritten with the name of the target (which comes from the user so it matches what they want rather than using the actual canonical name which might be different). 

This means that the nametable list has the same information in it as the destination list passed to a test, so name and address lookups in the test itself are much easier (and always possible, previously any addresses resolved at test time in measured had no name).