Add TTL and version fields to the icmp test report structures.
Version is set when generating results to make sure that any code interpreting the results is up to date - it would be easy for the server and client to end up on different versions of tests. 

TTL is still to be set in the icmp test, but pads out the result structure nicely.