Merge the two libraries libamp and libmeasured into a single one.
They were being kept separate so that xferd didn't end up with lots of code specific to running tests, but this is no longer an issue now that xferd is all done in python. This new combined library has code used by the tests, whether running stand-alone or built in.

There may now be stuff that can be moved out to measured rather than being in the shared library now that xferd is not trying to share it.