Replace the packet count/size schedule items with a total byte count option.
The test had a schedule option to control how many "packets" of a certain size were sent during a test (rather than a timed test). In actual fact this was simply the number of write calls to make, mostly unrelated to what TCP decides to do with packetisation. I've retained the ability to send a specific amount of data like the old schedule item count, but it is now measured in bytes.

Each test will use write sizes based on the global option and it won't be set per item in a test schedule.