Prevent late arriving responses from resetting missing response counters.
If a response packet arrived late it would reset the no_reply_count, which could lead to more than 5 timed out hops in a row on the end of the path. Also, the current TTL was being incremented on late response packets, which meant it was being incremented twice (as it had already been incremented when 2 queries had timed out and it moved on to the next TTL).