Add but disable some of the old code for checking maximum path lengths.
I'm still not 100% convinced this is required, but I wrote it while testing some unusual behaviour on VirtualBox, so might as well keep it. Theoretically it is possible for slow responding end hosts to be probed multiple times, causing the final responding hop to be later than it should, or have null hops following it.