Be smarter about resolving names - just those we can use or ask for.
As well as specifying the number of addresses a test should resolve, you can now also specify an address family if desired. It's not the nicest interface, but it shows how it can be done and at least exposes the functionality. This will need to be added to the standalone tests at some point.

The standalone tests should now use the AI_ADDRCONFIG flag like measured does to not resolve any addresses that the machine can not use (e.g. IPv6 without IPv6 connectivity).