Deal more gracefully in tests with packets that can't be sent.
If a machine has an IPv6 address but no routes, it will still resolve IPv6 addresses and try to test to them, but will fail to send packets. Previously this would result in the packets being marked as outstanding and so the code would wait for a response that would never arrive. Now, it should be able to fall through immediately and mark it as "done but failed/missing".