Accept source interface/addresses in the HTTP test but ignore it.
It isn't obvious the best way to get libcurl to use the information in the case where both a source IPv4 and IPv6 address are set (libcurl will just use the most recently set one). We probably need to only set the one that is relevant to the address family that will actually be used to fetch the page - do we need to try to figure out which family we will be using before we start?