Properly bounds check received data and short circuit harvest function.
Was possible for received data to be shorter than it claimed to be and so header fields would be indexed into invalid locations in memory. This should hopefully make sure that we have the minimum amount of data to read what we need.

Updated the harvest function to stop early once it knows that all outstanding packets have been received. This prevents the minimum 200ms wait that was happening, even with no outstanding packets.

Should move to git soon, so updates like these don't all get merged together!