Revert meaning of 'n' flag in throughput schedule
We had originally intended for 'n' to be only used in cases
where the connection was NOT reset between throughput runs (i.e.
the opposite to the current implementation) but editing the 
parsing code to allow this was too difficult. As a result, I've
inverted the parsing in the save function to match the original
behaviour rather than the proposed change.