Deal better with threaded resolver model (locking, worker assignment).
Each test thread will go into the waiting state once it has sent queries for all the names required, and of the threads waiting it is arbitrary which one will get any response. This means that accessing the addrlist pointer needs to be protected by a mutex and also means that tests need to know when all their destinations have been resolved. This last one is achieved through a shared counter variable, which when it reaches zero will cause the test thread to continue.