Set address family in sockaddr portion of addrinfo struct.
When converting results from the resolver into a list of addrinfo structs, the address family was only being set in the top level and not also in the sockaddr contained within it. Only the address itself was being set. Tests use the sockaddr as the destination for sending packets, which fails as it is not well formed.