Deal gracefully with test process going away during name resolution.
Previously if the test was killed for taking too long and there were still outstanding names to be resolved, it would eventually crash when it came time to report those names. The other end of the connection was not present which would trigger a SIGPIPE, which is now returned as an error code from send() and dealt with.