Fix various bugs in tcpping test found in testing
Most important change is that payload is NOT appended to the
TCP SYN by default, as payload-bearing SYNs are frequently dropped
by firewalls etc. The option to add payload is still available -- the
option itself has changed from a "packet size" to a "payload size"
option, which means the zero case will work for both IPv4 and IPv6.