amp-http: enlarge internal buffer for storing URLs
Fixes segfault in amp-http when run against www.amazon.com.

Increased internal buffer for storing paths to 2048 characters. Only the
first 256 characters are reported to ampsave to limit the result message
size.

Fixed a few other possible overruns and inconsistencies in the buffer
sizing in the HTTP test.

Updated ampsave for HTTP to handle the new buffer sizes.