Add "hidden" shortcut schedule option to tput test.
This makes life easier when trying to schedule tests through the web interface, as the modal currently needs a one-to-one mapping of input item and command line argument. The direction dropdown box sets this new "direction" argument that gets used to build a schedule in the absence of a specific schedule line.