Clear watchdog if test is killed by a non-SIGKILL.
If killed by a SIGKILL we can assume that we sent it, and already cleared the watchdog. If killed by some other signal then the watchdog is probably still running. If the user kills the test then we won't do anything - the watchdog will still expire and will log a message, but won't actually do anything useful.