Update RPM build scripts to work with current code.
The Makefile patch should hopefully be temporary - the version of python used here seems to be slightly broken with respect to distutils/setuptools. It works fine in Ubuntu with python-2.7.3-6deb7u2 and previously worked fine in Centos with python-2.6.6-37.

Don't really care as the files get deleted anyway for the client packages, just create the directory and let it do what it thinks it needs to do.