Ensure tests don't get redscheduled immediately.
If a test was successfully run and the next scheduled time for the test is in the immediate future, try to roll over to the next scheduled time instead. Assume that the clock is running slow and that the test just triggered is the one for the current test period.