Added option to HTTP test to force curl SSL version
This can be used to solve the 'BNZ' problem, which prevents the HTTP
test from running successfully if curl is left to try and figure
out the right SSL version on its own.

I'll leave it up to Brendon to decide if he wants to go with this
solution or if he has something better in mind.