Don't count failed objects in the global stats.
Any object that times out or fails to fetch for any reason will not update the global object count or last fetched times. This prevents the graphs from having plateaus at the timeout values, because they only plot the global stats (failed object fetches are still counted in the per server stats). This is not the ideal final solution - we should report counts of failed objects too so that they can be shown in the tooltip or similar.