Split report messages into blocks of 255 results.
Currently these tests use a uint8_t to store the total number of results in the message. We are starting to hit the limits of this and need to send more, but increasing the size of the variable means we have to change all the ampsave code too. This is all going to be replaced with Google protocol buffers "soon", so we can fix it properly then.