Remove unused stopset code from traceroute test and improve timers.
The stopset was never used because of issues with multiple upstreams making
some paths use the incorrect first couple of hops, and the code was a mess
so decided it was better off being removed.

As part of that refactor, tried to improve the use of timers for sending and
resending resending probes so that the loss timeout is based on the oldest
outstanding packet, and the send timeout triggers immediately if the inter
packet delay has already elapsed (e.g. while waiting for a response).