Fix segmentation fault when DNS query fails very badly.
If a poorly formatted name is given to the resolver then it will return a NULL result and set err to mark it as a syntax error. Make sure we don't use the result in this case. We also still need to decrement the counters used to track outstanding queries so that the wait loop will eventually terminate when there are no more queries.

Fixes #5.