Don't listen on all addresses for throughput/udpstream if one is set.
The code to serve tests expects only one of socket.socket and socket.socket6 to be set and ignores the other one. This can leave the unused socket listening on a port during the test, and then block the server from being re-established as it tries to bind it again to test the other direction.

If the server is started as part of the amplet2-client then it will automatically set the "-4" or "-6" flag with the address of the interface that should be used.

If the server is started standalone then the listen_control_server() function will tidy up after itself to make sure only one socket is configured.