Allow tests to be limited to a single address family.
Standalone tests can now take a -4 or -6 parameter with or without an address. If an address is specified then that is the source address used for targets in that address family (same behaviour as previously) and the other address family is disabled (new behaviour). If no address is given then it will force use of that address family on whichever address the operating system decides. You can combine both options to set an address for one or both address families.

The amplet2-client configuration can specify an address to use for each address family (same behaviour as previously), if only one is specified then the other is disabled (new behaviour). To keep one enabled without specifying an address, you can set it to "any".

Built in tests take the same parameters and behave the same way as standalone tests. These options will be in addition to any set in the client configuration.