Update YouTube test to present the standard test API (now forks itself).
Calling fork() and execve() means that the zygote can fork again as the
correct process, removing the need for prctl trickery to try to set the
correct link in /proc/self/exe and the need to run as a single process.
It also means conflicts between libssl and libboringssl are no longer a
problem as the test is an entirely new process without libssl loaded.