Fix result handing to correctly extract NSID responses.
Reporting NSID responses in the amp-server ampsave scripts has had to
change slightly to account for nntsc using the instance name as part
of the stream key - "instance" now reports the server name (as it
always has when instance was missing) and the new "nsid" field reports
the actual bytes received in the NSID response.

The nntsc behaviour might be what we ultimately want, but until that
is decided try not to break anything that is currently working.

Closes #24.