Make shovel prefetch-count configurable, setting it to 1 by default.
Not the perfect rate limit, but testing on the NZ AMP mesh has this
keeping incoming messages to an easily manageable level, while still
also emptying the queues at a reasonable rate. Higher latency between
client and collector might mean the default of 1 is too low, so it is
configurable.

Closes #38.