Allow apt-get update to fail if repository has no Release file.
Trying to build for a new distribution will fail as no packages have
yet been uploaded to the repository, so there is no Release file. We
don't mind - go ahead and update the other repositories, then upload
the built packages into the empty repository.

If it fails to run for an actual real reason and we needed packages
from those repositories then the build should fail anyway due to
missing dependencies etc.