Use the default version of clang where possible.
All recent releases of Debian and Ubuntu have clang/lld packages
that will build the youtube test, so use them. The llvm repository
doesn't provide packages for armhf, and it's become easier to have
a special case for xenial and stretch while getting out of the way
for all the other releases.